AudioUtils: Use tanf instead of tan
Although it will not make difference in modern C++ (since C++11),
specify tanf instead of tan for the prewarping calculation to
ensure floats throughout.

+++ plugins/mimid/Engine/AudioUtils.h
@@ -113,7 +113,7 @@ inline static float tptlpupw(float & state , float inp , float cutoff , float sr
 // TPT LPF w/ cutoff pre-warping
 inline static float tptlp(float& state, float inp, float cutoff, float srInv)
 {
-	cutoff = tan(cutoff * srInv * pi) ;
+	cutoff = tanf(cutoff * srInv * pi) ;
 	float v = (inp - state) * cutoff / (1 + cutoff);
 	float res = v + state;
 	state = res + v;
